% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acemogluetal}
\alias{acemogluetal}
\title{Acemoglu, et al. (2001) "The Colonial Origins of Comparative Development: An Empirical Investigation", AER, 91(5) December, pp. 1369-1401.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 64 rows and 8 columns.}
\usage{
acemogluetal
}
\description{
Acemoglu, et al. (2001) "The Colonial Origins of Comparative Development: An Empirical Investigation", AER, 91(5) December, pp. 1369-1401.
}
\keyword{datasets}
