% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_nlevels-function.R
\name{as_nlevels}
\alias{as_nlevels}
\title{Convert factor to numeric using level values}
\usage{
as_nlevels(x)
}
\arguments{
\item{x}{Factor to convert}
}
\description{
Converts a factor to numeric values based on the numeric interpretation of its levels.
}
