% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_bin-function.R
\name{auto_bin}
\alias{auto_bin}
\title{Apply binning to a variable using Weight of Evidence values}
\usage{
auto_bin(sbin, df = ds)
}
\arguments{
\item{sbin}{Binning information created by the bin() function}

\item{df}{Data frame to apply binning to (default: ds)}
}
\description{
Applies binning to a variable in a data frame using previously created binning information.
Replaces original values with Weight of Evidence (WoE) values.
}
