% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basque}
\alias{basque}
\title{Abadie, Alberto and Javier Gardeazabal (2003) "The Economic Costs of Conflict: A Case Study of the Basque Country", AER, 93(1), March, pp. 113-132}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 774 rows and 23 columns.}
\usage{
basque
}
\description{
Abadie, Alberto and Javier Gardeazabal (2003) "The Economic Costs of Conflict: A Case Study of the Basque Country", AER, 93(1), March, pp. 113-132
}
\keyword{datasets}
