% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-function.R
\name{bin}
\alias{bin}
\title{Create binning for a predictor variable}
\usage{
bin(
  x,
  y,
  data,
  type = "a",
  q = 0.2,
  cut = NULL,
  silent = F,
  custom_plot = T,
  custom_colour_scale_start = "#5720FF",
  custom_colour_scale_end = "grey"
)
}
\arguments{
\item{x}{Name of the predictor variable}

\item{y}{Name of the response variable}

\item{data}{Data frame containing the variables}

\item{type}{Binning type: 'a' (automatic), 'f' (factor), 'q' (quantile), 'c' (custom) (default: 'a')}

\item{q}{Quantile step for quantile binning (default: 0.2)}

\item{cut}{Vector of cut points for custom binning (optional)}

\item{silent}{Whether to suppress output (default: FALSE)}

\item{custom_plot}{Whether to use custom plot styling (default: TRUE)}

\item{custom_colour_scale_start}{Start color for custom plot gradient (default: '#5720FF')}

\item{custom_colour_scale_end}{End color for custom plot gradient (default: 'grey')}
}
\description{
Creates binning for a predictor variable with respect to a response variable.
Supports automatic, factor, quantile, and custom binning methods.
Can also produce Weight of Evidence plots.
}
