% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{card}
\alias{card}
\title{Card, David (1995) "Using Geographic Variation in College Proximity to Estimate the Return to Schooling", NBER Working Paper No. 4483}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3010 rows and 37 columns.}
\usage{
card
}
\description{
Card, David (1995) "Using Geographic Variation in College Proximity to Estimate the Return to Schooling", NBER Working Paper No. 4483
}
\keyword{datasets}
