% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{castle-doctrine-2000-2010}
\alias{castle-doctrine-2000-2010}
\title{Cheng, Cheng and Mark Hoekstra (2013) "Does Strengthening Self-Defense Law Deter Crime or Escalate Violence? Evidence from Expansions to Castle Doctrine" 48(3) pp. 821-854.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 550 rows and 41 columns.}
\usage{
castle-doctrine-2000-2010
}
\description{
Cheng, Cheng and Mark Hoekstra (2013) "Does Strengthening Self-Defense Law Deter Crime or Escalate Violence? Evidence from Expansions to Castle Doctrine" 48(3) pp. 821-854.
}
\keyword{datasets}
