% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{castle_doctrine_2000_2010}
\alias{castle_doctrine_2000_2010}
\title{Castle doctrine dataset}
\format{A data frame with 550 rows and 41 variables:
\describe{
  \item{state}{state id}
  \item{year}{castle doctrine law, =1 if CDL is effective}
  \item{sid}{leading indicator (0 to 2 years before the adoption of CDL)}
  \item{cdl}{=1 if states with case law classified as not requiring duty to retreat}
  \item{pre2_cdl}{=1 if CDL extends to any place one has a legal right to be}
  \item{caselaw}{=1 if CDL includes a presumption of reasonable fear}
  \item{anywhere}{=1 if CDL removes civil liability}
  \item{assumption}{homicide count}
  \item{civil}{proportion of robberies using gun}
  \item{homicide_c}{justifiable homicide by private citizen count}
  \item{robbery_gun_r}{justifiable homicide by police count}
  \item{jhcitizen_c}{homicide count per 100,000 state population}
  \item{jhpolice_c}{robbery count per 100,000 state population}
  \item{homicide}{aggravated assault count per 100,000 state population}
  \item{robbery}{burglary count per 100,000 state population}
  \item{assault}{larceny count per 100,000 state population}
  \item{burglary}{motor vehicle theft count per 100,000 state population}
  \item{larceny}{murder count per 100,000 state population}
  \item{motor}{felony and suspected Felony Type homicide count per 100,000 state population}
  \item{murder}{justifiable homicide by private citizen count per 100,000 state population}
  \item{hc_felonywsus}{justifiable homicide by police count per 100,000 state population}
  \item{jhcitizen}{state population}
  \item{jhpolice}{full-time equivalent police per 100,000 state population}
  \item{population}{unemployment rate}
  \item{police}{state median income}
  \item{unemployrt}{\% of black male ager 15-24}
  \item{income}{\% of white male ager 15-24}
  \item{blackm_15_24}{\% of black male ager 25-44}
  \item{whitem_15_24}{\% of white male ager 25-44}
  \item{blackm_25_44}{incarceration rate per 100,000 state population}
  \item{whitem_25_44}{lagged incarceration rate per 100,000 state population}
  \item{prisoner}{poverty rate}
  \item{lagprisoner}{government spending per capita on subsidy}
  \item{poverty}{government spending per capita on public welfare}
  \item{exp_subsidy}{=1 if northeast Census region}
  \item{exp_pubwelfare}{=1 if midwest Census region}
  \item{northeast}{=1 if south Census region}
  \item{midwest}{=1 if west Census region}
  \item{south}{CDL effective year}
  \item{west}{state id}
  \item{effyear}{castle doctrine law, =1 if CDL is effective}
}}
\source{
\url{http://jhr.uwpress.org/content/48/3/821.abstract}
}
\usage{
castle_doctrine_2000_2010
}
\description{
Data used in the Cheng and Hoekstra (2013) analysis of the effect of
castle doctrine gun laws on homicides.
}
\references{
Cheng Cheng and Mark Hoekstra. "Does Strengthening Self-Defense
Law Deter Crime or Escalate Violence? Evidence from Expansions to Castle
Doctrine." \emph{Journal of Human Resources} 48.3 (2013): 821-854.
}
\keyword{datasets}
