% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coalIV}
\alias{coalIV}
\title{Black, Dan A., Terra G. McKinnish and Seth G. Sanders (2003) "Does the Availability oof High-wage Jobs from Low-skilled Men Affect Welfare Expenditures? Evidence from Shocks to the Steel and Coal Industries", Journal of Public Economics, 87, 1921-1942}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8175 rows and 10 columns.}
\usage{
coalIV
}
\description{
Black, Dan A., Terra G. McKinnish and Seth G. Sanders (2003) "Does the Availability oof High-wage Jobs from Low-skilled Men Affect Welfare Expenditures? Evidence from Shocks to the Steel and Coal Industries", Journal of Public Economics, 87, 1921-1942
}
\keyword{datasets}
