% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cornwell}
\alias{cornwell}
\title{Cornwell, Christopher and William N. Trumbull (1994) "Estimating the Economic Model of Crime with Panel Data", Review of Economics and Statistics 76(2) May, 360-366.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 630 rows and 59 columns.}
\usage{
cornwell
}
\description{
Cornwell, Christopher and William N. Trumbull (1994) "Estimating the Economic Model of Crime with Panel Data", Review of Economics and Statistics 76(2) May, 360-366.
}
\keyword{datasets}
