% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cps_controls}
\alias{cps_controls}
\title{Dehejia, Rajeev H. and Sadek Wahba (2002) "Propensity Score-Matching Methods for Nonexperimental Causal Studies" Review of Economics and Statistics 84(1) February, 151-161.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15992 rows and 11 columns.}
\usage{
cps_controls
}
\description{
Dehejia, Rajeev H. and Sadek Wahba (2002) "Propensity Score-Matching Methods for Nonexperimental Causal Studies" Review of Economics and Statistics 84(1) February, 151-161.
}
\keyword{datasets}
