% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_file_location-function.R
\name{current_file_location}
\alias{current_file_location}
\title{Get the directory of the current script}
\usage{
current_file_location()
}
\description{
Determines the directory path of the currently executing R script.
Works in both RStudio and when running from command line.
}
