% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elapsed-functions.R
\name{elapsed_days}
\alias{elapsed_days}
\title{Calculate elapsed days between two dates}
\usage{
elapsed_days(start_date, end_date)
}
\arguments{
\item{start_date}{Start date}

\item{end_date}{End date}
}
\description{
Calculates the days between two dates.
}
