% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elapsed-functions.R
\name{elapsed_months}
\alias{elapsed_months}
\title{Calculate elapsed months between two dates}
\usage{
elapsed_months(start_date, end_date, accuracy_type = "day")
}
\arguments{
\item{start_date}{Start date}

\item{end_date}{End date}

\item{accuracy_type}{Type of accuracy calculation ('day' or 'sql') (default: 'day')}
}
\description{
Calculates the months between two dates using different accuracy types.
}
