% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elapsed-functions.R
\name{elapsed_weeks}
\alias{elapsed_weeks}
\title{Calculate elapsed weeks between two dates}
\usage{
elapsed_weeks(start_date, end_date, iso = F)
}
\arguments{
\item{start_date}{Start date}

\item{end_date}{End date}

\item{iso}{Whether to use ISO week calculation (default: FALSE)}
}
\description{
Calculates the weeks between two dates.
}
