% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish}
\alias{fish}
\title{Graddy, Kathryn "The Fulton Fish Market" Journal of Economic Perspectives, 20(2), Spring, 207-220. 1}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 97 rows and 11 columns.}
\usage{
fish
}
\description{
Graddy, Kathryn "The Fulton Fish Market" Journal of Economic Perspectives, 20(2), Spring, 207-220. 1
}
\keyword{datasets}
