% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix-functions.R
\name{\%bracket_max\%}
\alias{\%bracket_max\%}
\title{Find the next highest bracket value}
\usage{
x \%bracket_max\% range
}
\arguments{
\item{x}{Numeric value to bracket}

\item{range}{Numeric vector of bracket values}
}
\value{
The next highest bracket value from range
}
\description{
Returns the smallest value in range that is greater than or equal to x.
}
