% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix-functions.R
\name{\%bracket_min\%}
\alias{\%bracket_min\%}
\title{Find the next lowest bracket value}
\usage{
x \%bracket_min\% range
}
\arguments{
\item{x}{Numeric value to bracket}

\item{range}{Numeric vector of bracket values}
}
\value{
The next lowest bracket value from range
}
\description{
Returns the largest value in range that is less than or equal to x.
}
