% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix-functions.R
\name{\%> <\%}
\alias{\%> <\%}
\title{Check if value is within range (exclusive on both ends)}
\usage{
x \%> <\% range
}
\arguments{
\item{x}{Numeric value to test}

\item{range}{Numeric vector of length 2 defining the range}
}
\value{
Logical vector
}
\description{
Tests if x is greater than \code{range[1]} AND less than \code{range[2]}.
}
