% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix-functions.R
\name{\%limit\%}
\alias{\%limit\%}
\title{Limit values to a specified range}
\usage{
x \%limit\% range
}
\arguments{
\item{x}{Numeric value to limit}

\item{range}{Numeric vector of length 2 defining the min and max limits}
}
\value{
Numeric value constrained to the range
}
\description{
Constrains x to be within the specified range by capping values at \code{range[1]} (min) and \code{range[2]} (max).
}
