% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na-null-functions.R
\name{if_na}
\alias{if_na}
\title{Replace NA values in a vector}
\usage{
if_na(x, replacement, na = T, value = NULL)
}
\arguments{
\item{x}{Vector to process}

\item{replacement}{Value to use as replacement}

\item{na}{Whether to replace NA values (TRUE) or specific values (FALSE) (default: TRUE)}

\item{value}{Specific values to replace when na=FALSE}
}
\description{
Replaces NA values or specific values in a vector with a replacement value.
}
