% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na-null-functions.R
\name{if_null}
\alias{if_null}
\title{Replace NULL, infinite, or empty values in a vector}
\usage{
if_null(x, replacement = NA)
}
\arguments{
\item{x}{Vector to process}

\item{replacement}{Value to use as replacement (default: NA)}
}
\description{
Replaces NULL, infinite, empty strings, or special text values with a replacement value.
}
