% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_date-function.R
\name{is_date}
\alias{is_date}
\title{Check if an object is a date or datetime}
\usage{
is_date(x)
}
\arguments{
\item{x}{Object to check}
}
\description{
Checks if an object inherits from Date or POSIXt classes.
}
