% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift_chart-function.R
\name{lift_chart}
\alias{lift_chart}
\title{Create lift charts using tile-based binning}
\usage{
lift_chart(
  prob = "p_",
  y,
  measure = NULL,
  df_train,
  df_test,
  model = NULL,
  tile_breaks = NULL,
  n = 10,
  nudge_train = 1.5,
  nudge_test = 1.5,
  generate_output = T,
  actual_colour = mix_palette$blue,
  pred_colour = mix_palette$red,
  line_width = 1.1,
  point_size = 2.5,
  lift_bar_x = 10.5,
  lift_bar_colour = mix_palette$serene_purple
)
}
\arguments{
\item{prob}{Name of probability column (default: 'p_')}

\item{y}{Name of the target variable column}

\item{measure}{Name of measure column for additional analysis (optional)}

\item{df_train}{Training dataset}

\item{df_test}{Test dataset}

\item{model}{Model object (optional)}

\item{tile_breaks}{Pre-computed tile breaks to use instead of deriving from train data (default: NULL)}

\item{n}{Number of tiles (default: 10)}

\item{nudge_train}{Position adjustment for training labels (default: 1.5)}

\item{nudge_test}{Position adjustment for test labels (default: 1.5)}

\item{generate_output}{Whether to generate output (default: TRUE)}

\item{actual_colour}{Colour for actual values (default: mix_palette$blue)}

\item{pred_colour}{Colour for predicted values (default: mix_palette$red)}

\item{line_width}{Width of the actual/predict lines (default: 1.1)}

\item{point_size}{Size of the actual/predict points (default: 2.5)}

\item{lift_bar_x}{X position of the lift bar drawn to the right of the last tile (default: 10.5)}

\item{lift_bar_colour}{Colour of the lift bar and its label (default: mix_palette$serene_purple)}
}
\description{
Creates lift charts for model evaluation by binning predictions into tiles.
Works with training and test datasets and supports different model types.
}
