% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lmb_data}
\alias{lmb_data}
\title{Voting and ideology dataset}
\format{A data frame with 13588 rows and 178 variables:
\describe{
  \item{state}{ICPSR STATE CODE}
  \item{district}{DISTRICT CODE}
  \item{incmbncy}{INCUMBENCY CODE}
  \item{demvote}{NUMBER OF DEMOCRATIC VOTES}
  \item{repvote}{NUMBER OF REPUBLICAN VOTES}
  \item{year}{(mean) eq_Dwhip}
  \item{congress}{(mean) eq_Rwhip}
  \item{occupanc}{(mean) eq_Dlead}
  \item{name}{(mean) eq_Rlead}
  \item{eq_Dwhip}{(mean) vote}
  \item{eq_Rwhip}{(mean) republic}
  \item{eq_Dlead}{(mean) party}
  \item{eq_Rlead}{(mean) office}
  \item{vote}{(mean) id}
  \item{republic}{(mean) whip_D}
  \item{party}{(mean) whip_R}
  \item{office}{(count) id}
  \item{icpsr_id}{NAME}
  \item{whip_D}{P0010001}
  \item{whip_R}{P107A001}
  \item{vote_tot}{1}
  \item{demvoteshare}{2}
  \item{dembin}{0}
  \item{lagdemvoteshare}{1}
  \item{clusterid}{2}
  \item{pooleyear}{3}
  \item{redistrict}{4}
  \item{poolename}{5}
  \item{aclu_vs}{6}
  \item{acu_vs}{11}
  \item{aca_vs}{12}
  \item{ada_vs}{13}
  \item{afbf_vs}{14}
  \item{afge_vs}{21}
  \item{afscme_vs}{22}
  \item{aft_vs}{23}
  \item{asc_vs}{24}
  \item{bfw_vs}{25}
  \item{bctd_vs}{31}
  \item{carter_vs}{32}
  \item{ccus1_vs}{33}
  \item{ccus2_vs}{34}
  \item{cwla_vs}{35}
  \item{cv_vs}{36}
  \item{cvvf_vs}{37}
  \item{sane_vs}{40}
  \item{cfnfmp_vs}{41}
  \item{aflcio_vs}{42}
  \item{cfsca_vs}{43}
  \item{cfsce_vs}{44}
  \item{cfscd_vs}{45}
  \item{cfscs_vs}{46}
  \item{ccause_vs}{47}
  \item{cw_vs}{48}
  \item{cc_vs}{49}
  \item{cfa_vs}{51}
  \item{ike_vs}{52}
  \item{pfpike_vs}{53}
  \item{pdpike_vs}{54}
  \item{ford_vs}{55}
  \item{fcnl_vs}{56}
  \item{lbj_vs}{61}
  \item{pfplbj_vs}{62}
  \item{pdplbj_vs}{63}
  \item{jfk_vs}{64}
  \item{pfpjfk_vs}{65}
  \item{pdpjfk_vs}{66}
  \item{lfr_vs}{67}
  \item{lcv_vs}{68}
  \item{lwv_vs}{71}
  \item{ll_vs}{72}
  \item{lfs_vs}{73}
  \item{nasc_vs}{81}
  \item{ncsc_vs}{82}
  \item{nea_vs}{0}
  \item{nfo_vs}{1}
  \item{nfu_vs}{0}
  \item{nfib_vs}{1}
  \item{ntu_vs}{2}
  \item{nwpc_vs}{0}
  \item{nr_vs}{1}
  \item{nixon_vs}{2}
  \item{reagan_vs}{3}
  \item{ripon_vs}{0}
  \item{twr_vs}{1}
  \item{uaw_vs}{2}
  \item{umw_vs}{0}
  \item{firstyear}{1}
  \item{adayear}{2}
  \item{nomada}{3}
  \item{realada}{4}
  \item{demvs2}{5}
  \item{demvs3}{6}
  \item{demvs4}{group(state district year n)}
  \item{lagdemvs2}{group(state district)}
  \item{lagdemvs3}{ICPSR STATE CODE}
  \item{lagdemvs4}{DISTRICT CODE}
  \item{dem2}{INCUMBENCY CODE}
  \item{lagdem2}{NUMBER OF DEMOCRATIC VOTES}
  \item{lagaclu}{NUMBER OF REPUBLICAN VOTES}
  \item{lagacu}{(mean) eq_Dwhip}
  \item{lagaca}{(mean) eq_Rwhip}
  \item{lagafbf}{(mean) eq_Dlead}
  \item{lagafge}{(mean) eq_Rlead}
  \item{lagafscme}{(mean) vote}
  \item{lagaft}{(mean) republic}
  \item{lagasc}{(mean) party}
  \item{lagbfw}{(mean) office}
  \item{lagbctd}{(mean) id}
  \item{lagcarter}{(mean) whip_D}
  \item{lagccus1}{(mean) whip_R}
  \item{lagcwla}{(count) id}
  \item{lagcv_vs}{NAME}
  \item{lagcvvf}{P0010001}
  \item{lagsane}{P107A001}
  \item{lagcfnfmp}{1}
  \item{lagaflcio}{2}
  \item{lagcfsca}{0}
  \item{lagcfsce}{1}
  \item{lagcfscd}{2}
  \item{lagcfscs}{3}
  \item{lagccause}{4}
  \item{lagcw_vs}{5}
  \item{lagcc_vs}{6}
  \item{lagcfa}{11}
  \item{lagike}{12}
  \item{lagford}{13}
  \item{laglbj}{14}
  \item{laglfr}{21}
  \item{laglcv}{22}
  \item{laglwv}{23}
  \item{lagll}{24}
  \item{laglfs}{25}
  \item{lagnasc}{31}
  \item{lagncsc}{32}
  \item{lagnea}{33}
  \item{lagnfo}{34}
  \item{lagnfu}{35}
  \item{lagnfib}{36}
  \item{lagntu}{37}
  \item{lagnwpc}{40}
  \item{lagnr}{41}
  \item{lagnixon}{42}
  \item{lagreagan}{43}
  \item{lagripon}{44}
  \item{lagtwr}{45}
  \item{laguaw}{46}
  \item{lagumw}{47}
  \item{lagnomada}{48}
  \item{totpop}{49}
  \item{medianincome}{51}
  \item{pcturban}{52}
  \item{pctblack}{53}
  \item{mnfcng}{54}
  \item{pcthighschl}{55}
  \item{votingpop}{56}
  \item{realincome}{61}
  \item{avgada}{62}
  \item{sex}{63}
  \item{stateborn}{64}
  \item{relatives_shrt}{65}
  \item{secondaryeduc}{66}
  \item{collegeattend}{67}
  \item{militaryservice}{68}
  \item{lastoccup_shrt}{71}
  \item{yrsofserv}{72}
  \item{age}{73}
  \item{pres_dem_pct}{81}
  \item{pres_rep_pct}{82}
  \item{n}{0}
  \item{lagged}{1}
  \item{_merge}{0}
  \item{democrat}{1}
  \item{lagdemocrat}{2}
  \item{score}{0}
  \item{lagscore}{1}
  \item{id}{2}
  \item{id2}{3}
}}
\source{
\url{https://academic.oup.com/qje/article-abstract/119/3/807/1938834?redirectedFrom=fulltext}
}
\usage{
lmb_data
}
\description{
Data used in the Lee et al (2004) analysis of voting and ideology.
}
\references{
David S. Lee, Enrico Moretti, and Matthew J. Butler. "Do Voters
Affect or Elect Policies? Evidence from the U.S. House." \emph{The Quarterly
Journal of Economics} 119.3 (2004): 807-859.
}
\keyword{datasets}
