% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_blob_read_rds-function.R
\name{mix_blob_read_rds}
\alias{mix_blob_read_rds}
\title{Read a single RDS file from Azure Blob Storage}
\usage{
mix_blob_read_rds(storage_account_name, container_name, file_path, blob_sas)
}
\arguments{
\item{storage_account_name}{The name of the Azure storage account}

\item{container_name}{The name of the container in the storage account}

\item{file_path}{The path to the RDS file (must include .rds extension)}

\item{blob_sas}{The shared access signature for authentication}
}
\value{
The object stored in the RDS file
}
\description{
Read a single RDS file from Azure Blob Storage
}
