% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_cluster-functions.R
\name{mix_cluster_stop}
\alias{mix_cluster_stop}
\title{Stop parallel processing cluster}
\usage{
mix_cluster_stop()
}
\value{
Stops the cluster and prints uptime message
}
\description{
Terminates the active parallel processing cluster and reports the total uptime
if a cluster was previously created with mix_cluster_make.
}
