% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_gdr_read-function.R
\name{mix_gdr_read}
\alias{mix_gdr_read}
\title{Read a file from Google Drive}
\usage{
mix_gdr_read(file_id, read_fn)
}
\arguments{
\item{file_id}{Google Drive file ID}

\item{read_fn}{Function to use for reading the file}
}
\description{
Downloads and reads a file from Google Drive using a provided reader function.
Handles authentication and error conditions.
}
