% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_install_packages-function.R
\name{mix_install_packages}
\alias{mix_install_packages}
\title{Install common R packages}
\usage{
mix_install_packages(
  install_core_packages = T,
  exclude_from_core = NULL,
  install_additional_packages = T
)
}
\arguments{
\item{install_core_packages}{Whether to install core packages (default: TRUE)}

\item{exclude_from_core}{Vector of package names to exclude from core installation (optional)}

\item{install_additional_packages}{Whether to install additional specialized packages (default: TRUE)}
}
\description{
Installs a curated set of R packages for data science workflows.
Includes core packages and additional specialized packages.
}
