% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_load_packages-function.R
\name{mix_load_packages}
\alias{mix_load_packages}
\title{Load commonly used packages}
\usage{
mix_load_packages()
}
\description{
Loads a set of commonly used R packages for data science workflows.
Includes packages for data manipulation, cloud services, and date handling.
}
