% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_mode-function.R
\name{mix_mode}
\alias{mix_mode}
\title{Find the mode of a vector}
\usage{
mix_mode(x)
}
\arguments{
\item{x}{Vector to find the mode of}
}
\description{
Finds the most frequently occurring value(s) in a vector.
If multiple values occur with the same frequency, all are returned.
}
