% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_palette.R
\docType{data}
\name{mix_palette}
\alias{mix_palette}
\title{Custom color palette}
\format{
A list of hex color codes
}
\usage{
mix_palette
}
\description{
A curated list of colors for data visualization, including blue, red, green,
orange, purple, yellow, and various shades for consistent theming.
}
\keyword{datasets}
