% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_palette.R
\docType{data}
\name{mix_palette_cb_jp}
\alias{mix_palette_cb_jp}
\title{Colorblind-friendly palette}
\format{
A vector of 8 hex color codes
}
\usage{
mix_palette_cb_jp
}
\description{
A colorblind-friendly palette with 8 distinct colors optimized for accessibility.
}
\keyword{datasets}
