% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_palette.R
\docType{data}
\name{mix_palette_gg}
\alias{mix_palette_gg}
\title{ggplot2 theme color palette}
\format{
A list of hex color codes for grey shades
}
\usage{
mix_palette_gg
}
\description{
Color palette matching ggplot2 default grey scale theme colors.
}
\keyword{datasets}
