% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_palette.R
\docType{data}
\name{mix_palette_light}
\alias{mix_palette_light}
\title{Light color palette}
\format{
A list of hex color codes for light/pastel shades
}
\usage{
mix_palette_light
}
\description{
A collection of light, pastel colors for softer data visualizations.
}
\keyword{datasets}
