% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_train_test_split-function.R
\name{mix_train_test_split}
\alias{mix_train_test_split}
\title{Create and apply training/test split}
\usage{
mix_train_test_split(
  df = ds,
  target = "t_",
  p = 0.7,
  create_index = F,
  split_data = T,
  df_train = ds1,
  df_test = ds2,
  env_var = NULL
)
}
\arguments{
\item{df}{Data frame to split (default: ds)}

\item{target}{Name of the target variable column for stratified sampling (default: 't_')}

\item{p}{Proportion of data to use for training (default: 0.7)}

\item{create_index}{Whether to create a new train_index (TRUE) or use existing (FALSE) (default: TRUE)}

\item{split_data}{Whether to split into separate train/test dataframes (default: FALSE)}

\item{df_train}{Name for the training dataframe variable if split_data=TRUE (default: ds1)}

\item{df_test}{Name for the test dataframe variable if split_data=TRUE (default: ds2)}

\item{env_var}{Environment to assign variables to (default: global environment)}
}
\description{
Creates or applies a training/test split index using data.table for performance.
Can either create a new stratified split or use an existing train_index column.
Optionally splits the data into separate training and test datasets.
}
