% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nsw_dw}
\alias{nsw_dw}
\title{Dehejia, Rajeev H. and Sadek Wahba (2002) "Propensity Score-Matching Methods for Nonexperimental Causal Studies" Review of Economics and Statistics 84(1) February, 151-161. 1}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2675 rows and 16 columns.}
\source{
\url{http://www.uh.edu/~adkugler/Dehejia&Wahba.pdf}
}
\usage{
nsw_dw
}
\description{
Dehejia, Rajeev H. and Sadek Wahba (2002) "Propensity Score-Matching Methods for Nonexperimental Causal Studies" Review of Economics and Statistics 84(1) February, 151-161. 1
}
\keyword{datasets}
