% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_roc_plot-function.R
\name{pr_roc_plot}
\alias{pr_roc_plot}
\title{Create Precision-Recall curve plots for model evaluation}
\usage{
pr_roc_plot(
  prob = "p_",
  y,
  df_train,
  df_test = NULL,
  model = NULL,
  generate_output = T,
  train_colour = mix_palette$blue,
  test_colour = mix_palette$red
)
}
\arguments{
\item{prob}{Name of probability column (default: 'p_')}

\item{y}{Name of the target variable column}

\item{df_train}{Training dataset}

\item{df_test}{Test dataset (optional)}

\item{model}{Model object (optional)}

\item{generate_output}{Whether to generate output (default: TRUE)}

\item{train_colour}{Colour for training curve (default: mix_palette$blue)}

\item{test_colour}{Colour for test curve (default: mix_palette$red)}
}
\description{
Creates Precision-Recall curve plots for model evaluation on training and optionally test datasets.
Calculates and displays PR-AUC. Particularly suited to imbalanced classification problems
where ROC-AUC can be misleadingly optimistic.
}
