% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty-functions.R
\name{pretty_curr}
\alias{pretty_curr}
\title{Format numeric values as currency}
\usage{
pretty_curr(x, p = 0, curr = "")
}
\arguments{
\item{x}{Numeric vector to format}

\item{p}{Number of decimal places (default: 0)}

\item{curr}{Currency symbol (default: "")}
}
\description{
Formats numeric values as currency with thousand separators and specified decimal places.
}
