% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty-functions.R
\name{pretty_num}
\alias{pretty_num}
\title{Format numeric values with thousand separators}
\usage{
pretty_num(x, p = 1)
}
\arguments{
\item{x}{Numeric vector to format}

\item{p}{Number of decimal places (default: 1)}
}
\description{
Formats numeric values with thousand separators and specified decimal places.
}
