% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty-functions.R
\name{pretty_perc}
\alias{pretty_perc}
\title{Format numeric values as percentages}
\usage{
pretty_perc(x, p = 2)
}
\arguments{
\item{x}{Numeric vector to format (in decimal form, e.g., 0.75 for 75\%)}

\item{p}{Number of decimal places (default: 2)}
}
\description{
Formats numeric values as percentages with specified decimal places.
}
