% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_plot-function.R
\name{roc_plot}
\alias{roc_plot}
\title{Create ROC curve plots for model evaluation}
\usage{
roc_plot(
  prob = "p_",
  y = "t_Paid",
  df_train = ds1,
  df_test = ds2,
  model = NULL,
  model_var = "X1",
  generate_plot = T,
  train_colour = mix_palette$blue,
  test_colour = mix_palette$red
)
}
\arguments{
\item{prob}{Prefix for probability column (default: 'p_')}

\item{y}{Name of the target variable column (default: "t_Paid")}

\item{df_train}{Training dataset (default: ds1)}

\item{df_test}{Test dataset (default: ds2)}

\item{model}{Model object (optional)}

\item{model_var}{Model variable to use for predictions (default: 'X1')}

\item{generate_plot}{Whether to generate plots (default: TRUE)}

\item{train_colour}{Colour for training curve (default: mix_palette$blue)}

\item{test_colour}{Colour for test curve (default: mix_palette$red)}
}
\description{
Creates ROC curve plots for model evaluation on training and optionally test datasets.
Calculates and displays AUC and Gini coefficients.
}
