% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{star_sw}
\alias{star_sw}
\title{Estimates-of-Education-Production Krueger, Alan B. (1999) "Experiental Estimates of Education Production Functions", Quarterly Journal of Economics 114(2) May, 497-532. 1}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11598 rows and 40 columns.}
\usage{
star_sw
}
\description{
Estimates-of-Education-Production Krueger, Alan B. (1999) "Experiental Estimates of Education Production Functions", Quarterly Journal of Economics 114(2) May, 497-532. 1
}
\keyword{datasets}
