% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substr_right_left-function.R
\name{substr_left}
\alias{substr_left}
\title{Extract leftmost characters from a string}
\usage{
substr_left(x, n)
}
\arguments{
\item{x}{Character vector}

\item{n}{Number of characters to extract from the left}
}
\description{
Extracts the leftmost n characters from a string.
}
