% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substr_right_left-function.R
\name{substr_right}
\alias{substr_right}
\title{Extract rightmost characters from a string}
\usage{
substr_right(x, n)
}
\arguments{
\item{x}{Character vector}

\item{n}{Number of characters to extract from the right}
}
\description{
Extracts the rightmost n characters from a string.
}
