% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{texas}
\alias{texas}
\title{Balanced panel (1985-2000) of imprisonment by race and other covariates 1}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 816 rows and 24 columns.}
\usage{
texas
}
\description{
Balanced panel (1985-2000) of imprisonment by race and other covariates 1
}
\keyword{datasets}
