% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_plot-function.R
\name{tile_plot}
\alias{tile_plot}
\title{Create lift charts using tile-based binning}
\usage{
tile_plot(
  prob = "p_",
  y,
  measure = NULL,
  df_train = ds1,
  df_test = ds2,
  model = NULL,
  model_var = "X1",
  n = 10,
  nudge_train = 1.5,
  nudge_test = 1.5,
  generate_plot = T,
  actual_colour = mix_palette$blue,
  pred_colour = mix_palette$red
)
}
\arguments{
\item{prob}{Prefix for probability column (default: 'p_')}

\item{y}{Name of the target variable column}

\item{measure}{Name of measure column for additional analysis (optional)}

\item{df_train}{Training dataset (default: ds1)}

\item{df_test}{Test dataset (default: ds2)}

\item{model}{Model object (optional)}

\item{model_var}{Model variable to use for predictions (default: 'X1')}

\item{n}{Number of tiles (default: 10)}

\item{nudge_train}{Position adjustment for training labels (default: 1.5)}

\item{nudge_test}{Position adjustment for test labels (default: 1.5)}

\item{generate_plot}{Whether to generate plots (default: TRUE)}

\item{actual_colour}{Colour for actual values (default: mix_palette$blue)}

\item{pred_colour}{Colour for predicted values (default: mix_palette$red)}
}
\description{
Creates lift charts for model evaluation by binning predictions into tiles.
Works with training and test datasets and supports different model types.
}
