% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanic}
\alias{titanic}
\title{Titanic manifest, both survivors and those who perished along several dimensions (e.g., sex, age, passenger class) 2}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2201 rows and 4 columns.}
\usage{
titanic
}
\description{
Titanic manifest, both survivors and those who perished along several dimensions (e.g., sex, age, passenger class) 2
}
\keyword{datasets}
