% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanic3}
\alias{titanic3}
\title{Titanic manifest, both survivors and those who perished along several dimensions (e.g., sex, age, passenger class) 2}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1310 rows and 14 columns.}
\usage{
titanic3
}
\description{
Titanic manifest, both survivors and those who perished along several dimensions (e.g., sex, age, passenger class) 2
}
\keyword{datasets}
