% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ucr2005}
\alias{ucr2005}
\title{UCR and LEOKA files, 1960-2005, from FBI 1}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 96994 rows and 19 columns.}
\source{
\url{http://eml.berkeley.edu/~jmccrary/UCR/index.html}
}
\usage{
ucr2005
}
\description{
UCR and LEOKA files, 1960-2005, from FBI 1
}
\keyword{datasets}
