% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yule}
\alias{yule}
\title{Yule, George Udny "An Investigation into the Causes of Changes in Pauperism in England, Chiefly During the Last Two Interensal Decades", Journal of Royal Statistical Society, 62, 249-295}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 32 rows and 5 columns.}
\usage{
yule
}
\description{
Yule, George Udny "An Investigation into the Causes of Changes in Pauperism in England, Chiefly During the Last Two Interensal Decades", Journal of Royal Statistical Society, 62, 249-295
}
\keyword{datasets}
