% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acemogluetal}
\alias{acemogluetal}
\title{Colonial dataset}
\format{A data frame with 64 rows and 8 variables:
\describe{
  \item{countryn}{full country name}
  \item{shortnam}{3 letter country name}
  \item{africa}{dummy = 1 for Africa}
  \item{latitude}{abs(latitude of capital) / 90}
  \item{euro}{\% of European descent 1975, AJR}
  \item{prot}{prot against exprop, AJR}
  \item{lgdp}{Log GDP per cpaita}
  \item{logmort}{Log settler mortality}
}}
\source{
\url{http://economics.mit.edu/files/4123}
}
\usage{
acemogluetal
}
\description{
Data used in Acemoglu et al (2001) analysis of the effet of instiutions on
economic performance.
}
\references{
Daron Acemoglu, Simon Johnson, James A. Robinson. "The Colonial
Origins of Comparative Development: An Empirical Investigation."
\emph{American Economic Review} 91.5 (2001):1369-1401.
}
\seealso{
\code{\link{ajrcomment}}
}
\keyword{datasets}
