% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ajrcomment}
\alias{ajrcomment}
\title{Albouy, David Y. (2012)}
\format{A data frame with 64 rows and 43 variables:
\describe{
  \item{longname}{full country name}
  \item{shortnam}{3 letter country name}
  \item{step}{Step in data construction}
  \item{mort}{Original Settler Mortality}
  \item{mort2}{Alternative mortality rates}
  \item{logmort0}{Original Log Settler Mortality}
  \item{logmort2}{log(mort2)}
  \item{risk}{Exp Risk}
  \item{loggdp}{Log GDP per capita (PPP) in 1995}
  \item{campaign}{Original mortality from campaign}
  \item{campaign2}{Campaign indicator for mort2}
  \item{source0}{Country source of own m. rate}
  \item{source2}{Source of own rate, mort2}
  \item{slave}{Data from forced labor}
  \item{latitude}{Absolute value of latitude}
  \item{neoeuro}{'New-Europes'(USA, CAN, AUS, NZL)}
  \item{asia}{Asia}
  \item{africa}{Africa}
  \item{other}{'Other' Cont (AUS, NZL, MLT!)}
  \item{rainmin}{Minimum Monthly Rainfall}
  \item{meantemp}{Mean Temperature (Parker, 1997)}
  \item{edes1975}{Pop of Euro Descent, 1975}
  \item{malaria}{Malaria in 1994}
  \item{nafrica}{North Africa}
  \item{other2}{Other Continent with USA & CAN}
  \item{asiae_bl}{East Asia, Barro-X}
  \item{laam_bl}{Latin America, Barro-X}
  \item{oecd_bl}{OECD, Barro-X}
  \item{safrica_bl}{Subsaharan Africa, Barro-X}
  \item{col_fr}{Colony of France}
  \item{col_uk}{Colony of UK}
  \item{cons90}{Constraint on Executive, 1990}
  \item{lado1995}{Law and Order Tradition in 1995}
  \item{urb1500}{Urbanization in 1500}
  \item{pd1500}{Population Density in 1500}
  \item{bishops}{Data from Gutierrez (1986)}
  \item{epidemic}{Data from epidemic period}
  \item{notsoudan}{data from Mali}
  \item{ajr_ead}{AJR's 'Earliest Available Data'}
  \item{inconsis}{Inconsistency correction (mort2)}
  \item{new_data}{Rates from AJR Response (mort2)}
  \item{mort_bis}{Bishop mortality}
  \item{mort_set}{Actual settler mortality}"
}}
\source{
\url{https://pdfs.semanticscholar.org/f179/00569cfd18339b3dbc53e66e2718b5db4a9c.pdf}
}
\usage{
ajrcomment
}
\description{
Data from "The Colonial Origins of Comparative Development: An Empirical
Investigation: Comment", AER, 102(6), pp. 3059-3076
}
\keyword{datasets}
