% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ajrcomment}
\alias{ajrcomment}
\title{Albouy, David Y. (2012) "The Colonial Origins of Comparative Development: An Empirical Investigation: Comment", AER, 102(6), pp. 3059-3076}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 64 rows and 43 columns.}
\usage{
ajrcomment
}
\description{
Albouy, David Y. (2012) "The Colonial Origins of Comparative Development: An Empirical Investigation: Comment", AER, 102(6), pp. 3059-3076
}
\keyword{datasets}
