% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basque}
\alias{basque}
\title{Abadie, Alberto and Javier Gardeazabal (2003)}
\format{A data frame with 774 rows and 23 variables:
\describe{
   \item{regionno}{Region number}
   \item{regionname}{Region name (string)}
   \item{year}{Year}
   \item{gdpcap}{GDP per capita}
   \item{secagriculture}{Agriculture, forestry and fishing, sectoral shares (percentages)}
   \item{secenergy}{Energy and Water, sectoral shares (percentages)}
   \item{secindustry}{Industry, sectoral shares (percentages)}
   \item{secconstruction}{Construction and engineering, sectoral shares (percentages)}
   \item{secservicesventa}{Marketable services, sectoral shares (percentages)}
   \item{secservicesnonventa}{Nonmarketable services, sectoral shares (percentages)}
   \item{schoolillit}{Human capital, illiterates}
   \item{schoolprim}{Human capital, primary or without studies}
   \item{schoolmed}{Human capital, medium}
   \item{schoolhigh}{Human capital, high school}
   \item{schoolposthigh}{Human capital, post-high school}
   \item{popdens}{Population density}
   \item{invest}{Investment ratio (percentage)}
   \item{panel_id}{Region name (integer with labels)}
   \item{school_illit_share}{Human capital, illiterates (percentage shares)}
   \item{school_prim_share}{Human capital, primary or without studies (percentage shares)}
   \item{school_med_share}{Human capital, medium (percentage shares)}
   \item{school_high_share}{Human capital, high school (percentage shares)}
   \item{school_posthigh_share}{Human capital, post-high school (percentage shares)}
}}
\source{
\url{http://www.nyu.edu/gsas/dept/politics/faculty/beck/abadie_aer.pdf}
}
\usage{
basque
}
\description{
Data from "The Economic Costs of Conflict: A Case Study of the Basque
Country", AER, 93(1), March, pp. 113-132
}
\keyword{datasets}
