% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{card}
\alias{card}
\title{Schooling and earnings dataset}
\format{A data frame with 3010 rows and 37 variables:
\describe{
  \item{id}{Residuals}
  \item{nearc2}{Fitted values}
  \item{nearc4}{Residuals}
  \item{educ}{Residuals}
  \item{age}{Fitted values}
  \item{fatheduc}{Residuals}
  \item{motheduc}{Residuals}
  \item{weight}{Fitted values}
  \item{momdad14}{Residuals}
  \item{sinmom14}{Residuals}
  \item{step14}{Fitted values}
  \item{reg661}{Residuals}
  \item{reg662}{Residuals}
  \item{reg663}{Fitted values}
  \item{reg664}{Residuals}
  \item{reg665}{Residuals}
  \item{reg666}{Fitted values}
  \item{reg667}{Residuals}
  \item{reg668}{Residuals}
  \item{reg669}{Fitted values}
  \item{south66}{Residuals}
  \item{black}{Residuals}
  \item{smsa}{Fitted values}
  \item{south}{Residuals}
  \item{smsa66}{Residuals}
  \item{wage}{Fitted values}
  \item{enroll}{Residuals}
  \item{KWW}{Residuals}
  \item{IQ}{Fitted values}
  \item{married}{Residuals}
  \item{libcrd14}{Residuals}
  \item{exper}{Fitted values}
  \item{lwage}{Residuals}
  \item{expersq}{Residuals}
  \item{u_lwage}{Fitted values}
  \item{lwage_hat}{Residuals}
  \item{u_educ}{Residuals}
}}
\source{
\url{https://www.dropbox.com/s/31xrncpk6gmlns3/Card\%201995.pdf?dl=0}
}
\usage{
card
}
\description{
Data used in the Card (1995) analysis of schooling and earnings.
}
\references{
Card, David. "Using Geographic Variation in College Proximity to
Estimate the Return to Schooling." NBER Working Paper (1995): 4483.
}
\keyword{datasets}
