% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coalIV}
\alias{coalIV}
\title{Coal dataset}
\format{A data frame with 8175 rows and 10 variables:
\describe{
  \item{year}{million tons coal 20+ seam}
  \item{coalres}{population in thousands (000)}
  \item{pop}{million tons coal 20+ seam}
  \item{coalprice}{population in thousands (000)}
  \item{earn}{million tons coal 20+ seam}
  \item{cat}{population in thousands (000)}
  \item{t}{million tons coal 20+ seam}
  \item{state}{population in thousands (000)}
  \item{county}{million tons coal 20+ seam}
  \item{afdc}{population in thousands (000)}
}}
\source{
\url{https://www.sciencedirect.com/science/article/abs/pii/S0047272702000142}
}
\usage{
coalIV
}
\description{
Data used in the Black et al. (2003) analysis of the effect of changes in
demand for low-skilled workers on welfare expenditures.
}
\references{
Dan A. Black, Terra G. McKinnish and Seth G. Sanders. "Does the
Availability of High-wage Jobs from Low-skilled Men Affect Welfare
Expenditures? Evidence from Shocks to the Steel and Coal Industries."
\emph{Journal of Public Economics} 87 (2003): 1921-1942
}
\keyword{datasets}
