% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts-function.R
\name{dts}
\alias{dts}
\title{Split data into training and test sets}
\usage{
dts(df = ds, x = "train_index", df_train = ds1, df_test = ds2, env_var = NULL)
}
\arguments{
\item{df}{Data frame to split (default: ds)}

\item{x}{Name of the index column (default: "train_index")}

\item{df_train}{Name for the training data frame variable (default: ds1)}

\item{df_test}{Name for the test data frame variable (default: ds2)}

\item{env_var}{Environment to assign variables to (default: global environment)}
}
\description{
Splits a data frame into training and test sets based on an index column.
Assigns the resulting data frames to the specified variables in the environment.
}
