% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lalonde_1986}
\alias{lalonde_1986}
\title{LaLonde, Robert J. (1986)}
\format{A data frame with 772 rows and 10 variables:
\describe{
  \item{data_id}{}
  \item{treat}{}
  \item{age}{}
  \item{education}{}
  \item{black}{}
  \item{hispanic}{}
  \item{married}{}
  \item{nodegree}{}
  \item{re75}{}
  \item{re78}{}
}}
\source{
\url{http://people.hbs.edu/nashraf/LaLonde_1986.pdf}
}
\usage{
lalonde_1986
}
\description{
Data from "Evaluating the Econometric Evaluations of Training Programs with
Experimental Data" AER, 76(4) September, 604-620.
}
\keyword{datasets}
