% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leoka2004}
\alias{leoka2004}
\title{LEOKA dataset (2004)}
\format{A data frame with 20585 rows and 10 variables:
\describe{
  \item{ori}{}
  \item{group}{}
  \item{year}{}
  \item{pop}{}
  \item{agency_name}{}
  \item{m_emp}{}
  \item{f_emp}{}
  \item{emp}{}
  \item{sworn}{}
  \item{civil}{}
}}
\source{
\url{https://eml.berkeley.edu/~jmccrary/UCR/index.html}
}
\usage{
leoka2004
}
\description{
Data from the Law Enforcement Officers Killed and Assaulted Program (LEOKA)
(2004). Collected by the FBI.
}
\seealso{
\code{\link{leoka2005}}, \code{\link{ucr2004}}, \code{\link{ucr2005}}
}
\keyword{datasets}
