% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leoka2005}
\alias{leoka2005}
\title{LEOKA 2005}
\format{A data frame with 20739 rows and 10 variables:
\describe{
  \item{ori}{}
  \item{group}{}
  \item{year}{}
  \item{pop}{}
  \item{agency_name}{}
  \item{m_emp}{}
  \item{f_emp}{}
  \item{emp}{}
  \item{sworn}{}
  \item{civil}{}
}}
\source{
\url{https://eml.berkeley.edu/~jmccrary/UCR/index.html}
}
\usage{
leoka2005
}
\description{
Data from LEOKA files, 1960-2005, from FBI.
}
\keyword{datasets}
