% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{margin_victory}
\alias{margin_victory}
\title{Margin of victory dataset}
\format{A data frame with 1494 rows and 4 variables:
\describe{
  \item{state2}{(mean) demwon}
  \item{year2}{(mean) marginvvv}
  \item{demwon}{(mean) demwon}
  \item{marginvvv}{(mean) marginvvv}
}}
\source{
\url{https://www.aeaweb.org/articles?id=10.1257/app.20120387}
}
\usage{
margin_victory
}
\description{
Data used in the Beland (2015) analysis of party allegiance on labor market
outcomes.
}
\references{
Louis-Phillippe Beland. "Political Parties and Labor-Market
Outcomes: Evidence from US States." \emph{American Economic Journal: Applied
Economics} 7.4 (2015): 198-220.
}
\keyword{datasets}
