% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{margin_victory}
\alias{margin_victory}
\title{Beland, Louis-Phillippe (2015)}
\format{A data frame with 1494 rows and 4 variables:
\describe{
  \item{state2}{(mean) demwon}
  \item{year2}{(mean) marginvvv}
  \item{demwon}{(mean) demwon}
  \item{marginvvv}{(mean) marginvvv}
}}
\source{
\url{https://www.aeaweb.org/articles?id=10.1257/app.20120387}
}
\usage{
margin_victory
}
\description{
Data from "Political Parties and Labor-Market Outcomes: Evidence from US
States" AEJ: Applied, 7(4) October, 198-220.
}
\keyword{datasets}
