% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_azure_geo_read-function.R
\name{mix_azure_geo_read}
\alias{mix_azure_geo_read}
\title{Read a geo file from Azure Storage into an sf object}
\usage{
mix_azure_geo_read(
  storage_account_name,
  container_name,
  prefix,
  storage_key,
  storage_type = "adls",
  driver = "GeoJSON"
)
}
\arguments{
\item{storage_account_name}{Name of the Azure storage account}

\item{container_name}{Name of the container in the storage account}

\item{prefix}{Full path to the file within the container}

\item{storage_key}{Azure storage account key for authentication}

\item{storage_type}{Type of storage ('blob' or 'adls', default: 'adls')}

\item{driver}{OGR driver to use ('GeoJSON' or 'GPKG', default: 'GeoJSON')}
}
\value{
An sf object
}
\description{
Read a geo file from Azure Storage into an sf object
}
