% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_azure_geo_write-function.R
\name{mix_azure_geo_write}
\alias{mix_azure_geo_write}
\title{Write an sf object to Azure Storage as a geo file}
\usage{
mix_azure_geo_write(
  df,
  storage_account_name,
  container_name,
  prefix,
  storage_key,
  object_name,
  storage_type = "adls",
  driver = "GeoJSON"
)
}
\arguments{
\item{df}{sf object to write}

\item{storage_account_name}{Name of the Azure storage account}

\item{container_name}{Name of the container in the storage account}

\item{prefix}{Path to the destination folder within the container}

\item{storage_key}{Azure storage account key for authentication}

\item{object_name}{Base name for the output file (without extension)}

\item{storage_type}{Type of storage ('blob' or 'adls', default: 'adls')}

\item{driver}{OGR driver to use ('GeoJSON' or 'GPKG', default: 'GeoJSON')}
}
\description{
Write an sf object to Azure Storage as a geo file
}
