% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_azure_storage_list-function.R
\name{mix_azure_storage_list}
\alias{mix_azure_storage_list}
\title{List files in Azure Storage (Blob or ADLS)}
\usage{
mix_azure_storage_list(
  storage_account_name,
  container_name,
  prefix,
  storage_key,
  storage_type = "adls",
  object_format = NULL,
  regex_pattern = NULL
)
}
\arguments{
\item{storage_account_name}{Name of the Azure storage account}

\item{container_name}{Name of the container in the storage account}

\item{prefix}{Path to the folder within the container}

\item{storage_key}{Azure storage account key for authentication}

\item{storage_type}{Type of storage ('blob' or 'adls', default: 'adls')}

\item{object_format}{Format to filter files by ('parquet', 'csv', 'tsv', 'json', or NULL for all files) (default: NULL)}

\item{regex_pattern}{Optional regex pattern to further filter file names (default: NULL)}
}
\description{
Lists files in Azure Storage (Blob or ADLS), optionally filtering by file
type and/or a regex pattern.
}
