% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_train_test_split-functions.R
\name{mix_train_index}
\alias{mix_train_index}
\title{Create training index}
\usage{
mix_train_index(df, target = "t_", p = 0.7)
}
\arguments{
\item{df}{Data frame to add index to}

\item{target}{Name of the target variable column for stratified sampling (default: 't_')}

\item{p}{Proportion of data to use for training (default: 0.7)}
}
\description{
Creates a stratified train_index column using caret's createDataPartition.
}
