% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_train_test_split-functions.R
\name{mix_train_test_split}
\alias{mix_train_test_split}
\title{Split data into training and test sets}
\usage{
mix_train_test_split(df)
}
\arguments{
\item{df}{Data frame with train_index column}
}
\description{
Splits a data frame into training and test sets using the train_index column.
}
