% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nsw_dw}
\alias{nsw_dw}
\title{Dehejia, Rajeev H. and Sadek Wahba (2002)}
\format{A data frame with 2675 rows and 16 variables:
\describe{
  \item{data_id}{}
  \item{treat}{}
  \item{age}{}
  \item{educ}{}
  \item{black}{}
  \item{hisp}{}
  \item{marr}{}
  \item{nodegree}{}
  \item{re74}{}
  \item{re75}{}
  \item{re78}{}
  \item{age2}{}
  \item{educ2}{}
  \item{re742}{}
  \item{re752}{}
  \item{blacku74}{}
}}
\source{
\url{http://www.uh.edu/~adkugler/Dehejia&Wahba.pdf}
}
\usage{
nsw_dw
}
\description{
Data from "Propensity Score-Matching Methods for Nonexperimental Causal
Studies" Review of Economics and Statistics 84(1) February, 151-161.
}
\keyword{datasets}
