% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psid_controls}
\alias{psid_controls}
\title{Dehejia, Rajeev H. and Sadek Wahba (2002)}
\format{A data frame with 2490 rows and 11 variables:
\describe{
  \item{data_id}{}
  \item{treat}{}
  \item{age}{}
  \item{education}{}
  \item{black}{}
  \item{hispanic}{}
  \item{married}{}
  \item{nodegree}{}
  \item{re74}{}
  \item{re75}{}
  \item{re78}{}
}}
\source{
\url{http://www.uh.edu/~adkugler/Dehejia&Wahba.pdf}
}
\usage{
psid_controls
}
\description{
Data from "Propensity Score-Matching Methods for Nonexperimental Causal
Studies" Review of Economics and Statistics 84(1) February, 151-161.
}
\keyword{datasets}
