% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smoking}
\alias{smoking}
\title{Smoking dataset}
\format{A data frame with 1209 rows and 7 variables:
\describe{
  \item{state}{state no}
  \item{year}{year}
  \item{cigsale}{cigarette sale per capita (in packs)}
  \item{lnincome}{log state per capita gdp}
  \item{beer}{beer consumption per capita}
  \item{age15to24}{percent of state population aged 15-24 years}
  \item{retprice}{retail price of cigarettes}
}}
\source{
\url{http://economics.mit.edu/files/11859}
}
\usage{
smoking
}
\description{
Data used in the Abadie et al. (2010) analysis of a cigarette tax in
California called Proposition 99.
}
\references{
Alberto Abadie, Alexis Diamond, and Jens Hainmueller. "Synthetic
Control Methods for Comparative Case Studies: Estimating the Effect of
California's Tobacco Control Program." \emph{Journal of the American
Statistical Association} 105.490 (2010): 493-505.
}
\keyword{datasets}
