% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{star_sw}
\alias{star_sw}
\title{STAR dataset}
\format{A data frame with 11598 rows and 40 variables:
\describe{
  \item{stark}{attend project star class in kindergarten}
  \item{star1}{attend project star class in 1st grade}
  \item{star2}{attend project star class in 2nd grade}
  \item{star3}{attend project star class in 3rd grade}
  \item{totexpk}{years of total teaching experience - kindergarten}
  \item{treadssk}{total reading scaled score - sesat ii stanford}
  \item{tmathssk}{total math scaled score - sesat ii stanford}
  \item{totexp1}{years of total teaching experience - 1st grade}
  \item{treadss1}{total reading scaled score - primary 1 stanford}
  \item{tmathss1}{total math scaled score - primary 1 stanford}
  \item{totexp2}{years of total teaching experience - 2nd grade}
  \item{treadss2}{total reading scaled score - primary 2 stanford}
  \item{tmathss2}{total math scaled score - primary 2 stanford}
  \item{treadss3}{total reading scaled score - primary 3 stanford}
  \item{tmathss3}{total math scaled score - primary 3 stanford}
  \item{totexp3}{years of total teaching experience -  3rd grade}
  \item{schidkn}{school id-k (new) kindergarten}
  \item{schid1n}{school id-g1 (new) 1st grade}
  \item{schid2n}{school id-g2 (new) 2nd grade}
  \item{schid3n}{school id-g3 (new) 3rd grade}
  \item{tscorek}{attend project star class in kindergarten}
  \item{tscore1}{attend project star class in 1st grade}
  \item{tscore2}{attend project star class in 2nd grade}
  \item{tscore3}{attend project star class in 3rd grade}
  \item{sck}{years of total teaching experience - kindergarten}
  \item{sc1}{total reading scaled score - sesat ii stanford}
  \item{sc2}{total math scaled score - sesat ii stanford}
  \item{sc3}{years of total teaching experience - 1st grade}
  \item{rak}{total reading scaled score - primary 1 stanford}
  \item{ra1}{total math scaled score - primary 1 stanford}
  \item{ra2}{years of total teaching experience - 2nd grade}
  \item{ra3}{total reading scaled score - primary 2 stanford}
  \item{white}{total math scaled score - primary 2 stanford}
  \item{black}{total reading scaled score - primary 3 stanford}
  \item{other}{total math scaled score - primary 3 stanford}
  \item{boy}{years of total teaching experience -  3rd grade}
  \item{freelunk}{school id-k (new) kindergarten}
  \item{freelun1}{school id-g1 (new) 1st grade}
  \item{freelun2}{school id-g2 (new) 2nd grade}
  \item{freelun3}{school id-g3 (new) 3rd grade}
}}
\source{
\url{https://academic.oup.com/qje/article/114/2/497/1844226/Experimental-Estimates-of-Education-Production}
}
\usage{
star_sw
}
\description{
Data used in the Krueger (1999) analysis of the randomized experiment in
Tennessee called the Student/Teacher Achievement Ratio (STAR).
}
\references{
Krueger, Alan B. "Experiental Estimates of Education Production
Functions." \emph{Quarterly Journal of Economics} 114.2 (1999): 497-532.
}
\keyword{datasets}
