% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{texas}
\alias{texas}
\title{Texas dataset}
\format{A data frame with 816 rows and 24 variables:
\describe{
  \item{statefip}{Year}
  \item{year}{BM Prison}
  \item{race}{WM Prison}
  \item{bmprison}{BM POP}
  \item{wmprison}{WM POP}
  \item{bmpop}{Crack index}
  \item{wmpop}{AIDS mortality per 100,000 in t}
  \item{error}{parole}
  \item{crack}{probation}
  \item{alcohol}{capacity_rated}
  \item{income}{capacity_operational}
  \item{ur}{capacity_design}
  \item{poverty}{Year}
  \item{black}{BM Prison}
  \item{perc1519}{WM Prison}
  \item{aidscapita}{BM POP}
  \item{state}{WM POP}
  \item{parole}{Crack index}
  \item{probation}{AIDS mortality per 100,000 in t}
  \item{capacity_rated}{parole}
  \item{capacity_operational}{probation}
  \item{capacity_design}{capacity_rated}
  \item{bmprate}{capacity_operational}
  \item{wmprate}{capacity_design}
}}
\source{
\url{}
}
\usage{
texas
}
\description{
Balanced panel of imprisonment by race and other covariates (1985-2000).
}
\keyword{datasets}
