% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanic}
\alias{titanic}
\title{Titanic dataset}
\format{A data frame with 2201 rows and 4 variables:
\describe{
  \item{class}{class (ticket)}
  \item{age}{Age (Child vs Adult)}
  \item{sex}{Gender}
  \item{survived}{Survived}
}}
\source{
\url{https://www.dropbox.com/s/fxuph22p88vkacc/titanic3_readme.pdf?dl=0}
}
\usage{
titanic
}
\description{
Titanic manifest, both survivors and those who perished along several
dimensions (e.g., sex, age, passenger class).
}
\keyword{datasets}
