% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanic3}
\alias{titanic3}
\title{Titanic dataset}
\format{A data frame with 1310 rows and 14 variables:
\describe{
  \item{pclass}{pclass}
  \item{survived}{survived}
  \item{name}{name}
  \item{sex}{sex}
  \item{age}{age}
  \item{sibsp}{sibsp}
  \item{parch}{parch}
  \item{ticket}{ticket}
  \item{fare}{fare}
  \item{cabin}{cabin}
  \item{embarked}{embarked}
  \item{boat}{boat}
  \item{body}{body}
  \item{homedest}{home.dest}
}}
\source{
\url{https://www.dropbox.com/s/fxuph22p88vkacc/titanic3_readme.pdf?dl=0}
}
\usage{
titanic3
}
\description{
Titanic manifest, both survivors and those who perished along several
dimensions (e.g., sex, age, passenger class).
}
\keyword{datasets}
