% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ucr2005}
\alias{ucr2005}
\title{UCR 2005}
\format{A data frame with 96994 rows and 19 variables:
\describe{
  \item{YEAR}{YEAR}
  \item{AGENCY}{AGENCY NAME}
  \item{STNAME}{STATE NAME}
  \item{MONTH}{MONTH}
  \item{male_pro}{(sum) male_pro}
  \item{female_pro}{(sum) female_pro}
  \item{asian_pro}{(sum) asian_pro}
  \item{white_pro}{(sum) white_pro}
  \item{black_pro}{(sum) black_pro}
  \item{male_gamble}{(sum) male_gamble}
  \item{female_gamble}{(sum) female_gamble}
  \item{male_gta}{(sum) male_gta}
  \item{female_gta}{(sum) female_gta}
  \item{male_drug}{(sum) male_drug}
  \item{female_drug}{(sum) female_drug}
  \item{male_vandal}{(sum) male_vandal}
  \item{female_vandal}{(sum) female_vandal}
  \item{POP}{(max) POP}
  \item{date}{YEAR}
}}
\source{
\url{https://eml.berkeley.edu/~jmccrary/UCR/index.html}
}
\usage{
ucr2005
}
\description{
Data from UCR files, 2005, from FBI.
}
\keyword{datasets}
