% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yule}
\alias{yule}
\title{Paupers dataset}
\format{A data frame with 599 rows and 103 variables:
\describe{
  \item{ID}{ID}
  \item{Union}{Union}
  \item{Region}{Region}
  \item{County}{County}
  \item{Type}{Yule's Rural/Urban Classification of Unions}
  \item{BoothGroup}{Booth's Classification of Unions}
  \item{BoothDensity}{BoothDensity (People per Acre)}
  \item{Popn1871_2}{Population 1871 (from 1881 Pauperism Returns)}
  \item{MABI71}{Male Able-Bodied In-Door Paupers 1871}
  \item{FABI71}{Female Able-Bodied In-Door Paupers 1871}
  \item{ChABI71}{Child Able-Bodied In-Door Paupers 1871}
  \item{MNotABI71}{Male Not Able-Bodied In-Door Paupers 1871}
  \item{FNotABI71}{Female Not Able-Bodied In-Door Paupers 1871}
  \item{ChNotABI71}{Child Not Able-Bodied In-Door Paupers 1871}
  \item{MABO71}{Male Able-Bodied Out-Door Paupers 1871}
  \item{FABO71}{Female Able-Bodied Out-Door Paupers 1871}
  \item{ChABO71}{Child Able-Bodied Out-Door Paupers 1871}
  \item{MNotABO71}{Male Not Able-Bodied Out-Door Paupers 1871}
  \item{FNotABO71}{Female Not Able-Bodied Out-Door Paupers 1871}
  \item{ChNotABO71}{Child Not Able-Bodied Out-Door Paupers 1871}
  \item{MABI81}{Male Able-Bodied In-Door Paupers 1881}
  \item{FABI81}{Female Able-Bodied In-Door Paupers 1881}
  \item{ChABI81}{Child Able-Bodied In-Door Paupers 1881}
  \item{MNotABI81}{Male Not Able-Bodied In-Door Paupers 1881}
  \item{FNotABI81}{Female Not Able-Bodied In-Door Paupers 1881}
  \item{ChNotABI81}{Child Not Able-Bodied In-Door Paupers 1881}
  \item{MABO81}{Male Able-Bodied Out-Door Paupers 1881}
  \item{FABO81}{Female Able-Bodied Out-Door Paupers 1881}
  \item{ChABO81}{Child Able-Bodied Out-Door Paupers 1881}
  \item{MNotABO81}{Male Not Able-Bodied Out-Door Paupers 1881}
  \item{FNotABO81}{Female Not Able-Bodied Out-Door Paupers 1881}
  \item{ChNotABO81}{Child Not Able-Bodied Out-Door Paupers 1881}
  \item{Popn1881}{Population 1881 (from 1881 Census Population Abstracts)}
  \item{popratio7181}{Ratio of Population in 1881 to Population in 1871}
  \item{pauper71}{Proportion of Paupers in the Population in 1871}
  \item{pauper81}{Proportion of Paupers in the Population in 1881}
  \item{paupratio7181}{Ratio of 1881 Pauperism Proportion to 1871 Pauperism Proportion}
  \item{outratio71}{Ratio of Out-Door Paupers to In-Door Paupers 1871}
  \item{outratio81}{Ratio of Out-Door Paupers to In-Door Paupers 1881}
  \item{outratio7181}{Ratio of outratio81 to outratio71}
  \item{MABI91}{Male Able-Bodied In-Door Paupers 1891}
  \item{FABI91}{Female Able-Bodied In-Door Paupers 1891}
  \item{ChABI91}{Child Able-Bodied In-Door Paupers 1891}
  \item{MNotABI91}{Male Not Able-Bodied In-Door Paupers 1891}
  \item{FNotABI91}{Female Not Able-Bodied In-Door Paupers 1891}
  \item{ChNotABI91}{Child Not Able-Bodied In-Door Paupers 1891}
  \item{MABO91}{Male Able-Bodied Out-Door Paupers 1891}
  \item{FABO91}{Female Able-Bodied Out-Door Paupers 1891}
  \item{ChABO91}{Child Able-Bodied Out-Door Paupers 1891}
  \item{MNotABO91}{Male Not Able-Bodied Out-Door Paupers 1891}
  \item{FNotABO91}{Female Not Able-Bodied Out-Door Paupers 1891}
  \item{ChNotABO91}{Child Not Able-Bodied Out-Door Paupers 1891}
  \item{Popn1881_2}{Population 1881 (from 1891 Pauperism Returns)}
  \item{Popn1891}{Population 1891 (from 1891 Census Population Abstracts)}
  \item{popratio8191}{Ratio of Population in 1891 to Population in 1881}
  \item{pauper91}{Proportion of Paupers in the Population in 1891}
  \item{pauper81_2}{Proportion of Paupers in the Population in 1881 (using population numbers from 1}
  \item{paupratio8191}{Ratio of 1891 Pauperism Proportion to 1881 Pauperism Proportion}
  \item{outratio91}{Ratio of Out-Door Paupers to In-Door Paupers 1891}
  \item{outratio8191}{Ratio of outratio91 to outratio81}
  \item{mfab71I}{Ratio of Male to Female Able-Bodied In-Door Paupers in 1871}
  \item{mfab81I}{Ratio of Male to Female Able-Bodied In-Door Paupers in 1881}
  \item{mfab91I}{Ratio of Male to Female Able-Bodied In-Door Paupers in 1891}
  \item{mfab71O}{Ratio of Male to Female Able-Bodied Out-Door Paupers in 1871}
  \item{mfab81O}{Ratio of Male to Female Able-Bodied Out-Door Paupers in 1881}
  \item{mfab91O}{Ratio of Male to Female Able-Bodied Out-Door Paupers in 1891}
  \item{mfnotab71I}{Ratio of Male to Female Not Able-Bodied In-Door Paupers in 1871}
  \item{mfnotab81I}{Ratio of Male to Female Not Able-Bodied In-Door Paupers in 1881}
  \item{mfnotab91I}{Ratio of Male to Female Not Able-Bodied In-Door Paupers in 1891}
  \item{mfnotab71O}{Ratio of Male to Female Not Able-Bodied Out-Door Paupers in 1871}
  \item{mfnotab81O}{Ratio of Male to Female Not Able-Bodied Out-Door Paupers in 1881}
  \item{mfnotab91O}{Ratio of Male to Female Not Able-Bodied Out-Door Paupers in 1891}
  \item{adultchab71I}{Ratio of Adult to Child Able-Bodied In-Door Paupers 1871}
  \item{adultchab81I}{Ratio of Adult to Child Able-Bodied In-Door Paupers 1881}
  \item{adultchab91I}{Ratio of Adult to Child Able-Bodied In-Door Paupers 1891}
  \item{adultchnotab71I}{Ratio of Adult to Child Not Able-Bodied In-Door Paupers 1871}
  \item{adultchnotab81I}{Ratio of Adult to Child Not Able-Bodied In-Door Paupers 1881}
  \item{adultchnotab91I}{Ratio of Adult to Child Not Able-Bodied In-Door Paupers 1891}
  \item{adultchab71O}{Ratio of Adult to Child Able-Bodied Out-Door Paupers 1871}
  \item{adultchab81O}{Ratio of Adult to Child Able-Bodied Out-Door Paupers 1881}
  \item{adultchab91O}{Ratio of Adult to Child Able-Bodied Out-Door Paupers 1891}
  \item{adultchnotab71O}{Ratio of Adult to Child Not Able-Bodied Out-Door Paupers 1871}
  \item{adultchnotab81O}{Ratio of Adult to Child Not Able-Bodied Out-Door Paupers 1881}
  \item{adultchnotab91O}{Ratio of Adult to Child Not Able-Bodied Out-Door Paupers 1891}
  \item{Popn1871}{Population 1871 (from 1871 Census Population Abstracts)}
  \item{PopnM1871}{Male Population in 1871}
  \item{PopnF1871}{Female Population in 1871}
  \item{M6571}{Males Aged 65+ in the 1871 Population}
  \item{PropM6571}{Proportion of Males Aged 65+ in the 1871 Male Population}
  \item{F6571}{Females Aged 65+ in the 1871 Population}
  \item{PropF6571}{Proportion of Females Aged 65+ in the 1871 Female Population}
  \item{PopnM1881}{Male Population in 1881}
  \item{PopnF1881}{Female Population in 1881}
  \item{M6581}{Males Aged 65+ in the 1881 Population}
  \item{PropM6581}{Proportion of Males Aged 65+ in the 1881 Male Population}
  \item{F6581}{Females Aged 65+ in the 1881 Population}
  \item{PropF6581}{Proportion of Females Aged 65+ in the 1881 Female Population}
  \item{PopnM1891}{Male Population in 1891}
  \item{PopnF1891}{Female Population in 1891}
  \item{M6591}{Males Aged 65+ in the 1891 Population}
  \item{PropM6591}{Proportion of Males Aged 65+ in the 1891 Male Population}
  \item{F6591}{Females Aged 65+ in the 1891 Population}
  \item{PropF6591}{Proportion of Females Aged 65+ in the 1891 Female Population}
}}
\source{
\url{https://edisciplinas.usp.br/mod/resource/view.php?id=1418367}
}
\usage{
yule
}
\description{
Data used in the Yule (1899) analysis of public assistance and the number of
paupers.
}
\references{
Yule, George Udny. "An Investigation into the Causes of Changes
in Pauperism in England, Chiefly During the Last Two Interensal Decades."
\emph{"Journal of Royal Statistical Society} 62 (1899): 249-295.
}
\keyword{datasets}
